+++ src/main/java/net/kozibrodka/mocreatures/fuelsystem/ContainerHorseFuel.java
@@ -15,7 +15,7 @@ public class ContainerHorseFuel extends ScreenHandler
     {
         byte var3 = 51;
         horse = entityhorse;
-        addSlot(new Slot(entityhorse, 0, 80, 22)); //TODO SLOT NIEWPUSZCZAJACY
+        addSlot(new SlotHorseFuel(entityhorse, 0, 80, 22));
 
         for(int i1 = 0; i1 < 3; i1++)
         {